static cast fix

+++ Source/Aura/Private/UI/WidgetController/OverlayWidgetController.cpp
@@ -127,7 +127,7 @@ void UOverlayWidgetController::OnXPChanged(int32 NewXP) const
 		const int32 DeltaLevelUpRequirement = LevelUpRequirement - PreviousLevelUpRequirement;
 		const int32 XPForThisLevel = NewXP - PreviousLevelUpRequirement;
 
-		const float XPBarPercent = static_cast<float>(XPForThisLevel / DeltaLevelUpRequirement);
+		const float XPBarPercent = static_cast<float>(XPForThisLevel) / static_cast<float>(DeltaLevelUpRequirement);
 
 		OnXPPercentChangedDelegate.Broadcast(XPBarPercent);
 	}
